gomebrew: Move repeated logic from homebrew modules into module_utils
Fixes #8323.
